skills/grade/pyproject.toml
[project]
name = "clawaes"
version = "0.1.0"
description = "Score and auto-select the best images from a folder using CLIP"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "pillow",
    "numpy",
]
